requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
langchain==0.0.300
huggingface_hub==0.17.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic==2.4.2
boto3==1.28.57
mangum==0.17.0
openai>=1.0.0
python-dotenv>=1.0.0